Move custom elements to a folder
