+++ src/components/Link.astro
@@ -79,8 +79,8 @@ if (!urlIsProtected && plainUrl) {
 </style>
 
 <script>
-  import setupProtectedLinkWrapper from "../scripts/ProtectedLinkWrapper";
-  import setupPlainLinkWrapper from "../scripts/PlainLinkWrapper";
+  import setupProtectedLinkWrapper from "../scripts/Elements/ProtectedLinkWrapper";
+  import setupPlainLinkWrapper from "../scripts/Elements/PlainLinkWrapper";
   setupProtectedLinkWrapper();
   setupPlainLinkWrapper();
 </script>
