+++ src/components/Resume/ResumeExperience.astro
@@ -148,6 +148,6 @@ const startedRendered = formatDateTime(
 </style>
 
 <script>
-  import setupTimeDifferenceWrapper from "../../scripts/TimeDifferenceWrapper";
+  import setupTimeDifferenceWrapper from "../../scripts/Elements/TimeDifferenceWrapper";
   setupTimeDifferenceWrapper();
 </script>
